+++ common/src/main/java/net/spell_engine/client/SpellEngineClient.java
@@ -13,6 +13,7 @@ import net.spell_engine.client.render.SpellBindingBlockEntityRenderer;
 import net.spell_engine.config.ClientConfig;
 import net.spell_engine.config.ClientConfigWrapper;
 import net.spell_engine.config.HudConfig;
+import net.spell_engine.config.Tutorial;
 import net.spell_engine.spellbinding.SpellBindingBlockEntity;
 import net.spell_engine.spellbinding.SpellBindingScreen;
 import net.spell_engine.spellbinding.SpellBindingScreenHandler;
@@ -27,11 +28,18 @@ public class SpellEngineClient {
             .setDirectory(SpellEngineMod.ID)
             .sanitize(true)
             .build();
+    public static ConfigManager<Tutorial> tutorial = new ConfigManager<Tutorial>
+            ("spell_engine_tutorial", new Tutorial())
+            .builder()
+            .setDirectory("..") // One directory above config, to avoid file redistribution
+            .sanitize(true)
+            .build();
 
     public static void initialize() {
         AutoConfig.register(ClientConfigWrapper.class, PartitioningSerializer.wrap(JanksonConfigSerializer::new));
         config = AutoConfig.getConfigHolder(ClientConfigWrapper.class).getConfig().client;
         hudConfig.refresh();
+        tutorial.refresh();
 
         ClientNetwork.initializeHandlers();