+++ common/src/main/java/net/spell_engine/config/ClientConfig.java
@@ -32,9 +32,9 @@ public class ClientConfig implements ConfigData {
     @ConfigEntry.Gui.Tooltip
     public boolean showFocusedHotbarOnly = false;
     @ConfigEntry.Gui.Tooltip
-    public boolean lockHotbarOnRightClick = true;
+    public boolean lockHotbarOnRightClick = false;
     @ConfigEntry.Gui.Tooltip
-    public boolean unlockHotbarOnEscape = true;
+    public boolean unlockHotbarOnEscape = false;
     @ConfigEntry.Gui.Tooltip
     public boolean alwaysShowFullTooltip = false;
     @ConfigEntry.Gui.Tooltip