+++ common/src/main/java/net/spell_engine/config/Tutorial.java
@@ -1,0 +1,6 @@
+package net.spell_engine.config;
+
+public class Tutorial {
+    public String comment = "This file is not amongst other config files to avoid redistribution, so all users see the tutorial.";
+    public boolean spell_hotbar_shown = false;
+}